move bn stuff to initializers
