+++ examples/mlperf/initializers.py
@@ -3,6 +3,8 @@ from typing import Union, Tuple
 
 from tinygrad import Tensor, nn, dtypes
 from tinygrad.helpers import prod, argfix
+from tinygrad.multi import MultiLazyBuffer
+from examples.hlb_cifar10 import UnsyncedBatchNorm
 
 # rejection sampling truncated randn
 def rand_truncn(*shape, dtype=None, truncstds=2, **kwargs) -> Tensor:
@@ -78,3 +80,29 @@ class LayerNormBert:
     xn = x.cast(dtypes.float32).layernorm(eps=self.eps, axis=self.axis).cast(x.dtype)
     if not self.elementwise_affine: return xn
     return (xn * self.weight.cast(dtypes.default_float) + self.bias.cast(dtypes.default_float))
+
+class BatchNormRetinanet:
+  def __init__(self, sz:int, eps=1e-5, affine=True, track_running_stats=True, momentum=0.1):
+    self.eps, self.track_running_stats, self.momentum = eps, track_running_stats, momentum
+    if affine: self.weight, self.bias = Tensor.ones(sz, dtype=dtypes.float32), Tensor.zeros(sz, dtype=dtypes.float32)
+    else: self.weight, self.bias = None, None
+
+    self.running_mean, self.running_var = Tensor.zeros(sz, requires_grad=False, dtype=dtypes.float32), Tensor.ones(sz, requires_grad=False, dtype=dtypes.float32)
+    self.num_batches_tracked = Tensor.zeros(1, requires_grad=False, dtype=dtypes.int)
+  def __call__(self, x:Tensor):
+    batch_mean = self.running_mean
+    batch_invstd = self.running_var.reshape(1, -1, 1, 1).expand(x.shape).add(self.eps).rsqrt()
+    return x.batchnorm(self.weight, self.bias, batch_mean, batch_invstd).cast(dtypes.default_float)
+
+class UnSyncedBatchNormRetinanet(UnsyncedBatchNorm):
+  def __call__(self, x:Tensor):
+    if isinstance(x.lazydata, MultiLazyBuffer): assert x.lazydata.axis is None or x.lazydata.axis == 0 and len(x.lazydata.lbs) == self.num_devices
+    nd = x.shape[0]%self.num_devices
+    if nd==0: nd = self.num_devices
+    xr = x.reshape(nd, -1, *x.shape[1:]).cast(dtypes.float32)
+    batch_mean, batch_invstd = self.calc_stats(xr)
+    ret = xr.batchnorm(
+      self.weight.reshape(1, -1).expand((nd, -1)),
+      self.bias.reshape(1, -1).expand((nd, -1)),
+      batch_mean, batch_invstd, axis=(0, 2))
+    return ret.reshape(x.shape).cast(x.dtype)
