+++ examples/mlperf/model_train.py
@@ -378,9 +378,7 @@ def train_retinanet():
 
   from extra.models import retinanet
   from extra.models import resnet
-  from examples.hlb_cifar10 import UnsyncedBatchNorm
-  from tinygrad.multi import MultiLazyBuffer
-  from examples.mlperf.initializers import Conv2dNormal, Conv2dKaiming, Linear, Conv2dHeNormal
+  from examples.mlperf.initializers import Conv2dNormal, Conv2dKaiming, Linear, Conv2dHeNormal, BatchNormRetinanet, UnSyncedBatchNormRetinanet
   from tinygrad.nn.optim import Adam
   from extra.datasets.openimages import openimages, get_retinanet_train_files, get_retinanet_val_files
   from examples.mlperf.dataloader import batch_load_retinanet
@@ -397,34 +395,8 @@ def train_retinanet():
 
   resnet.Conv2d = Conv2dHeNormal
   resnet.Linear = Linear
-  if SYNCBN:
-    class CustBatchNorm:
-      def __init__(self, sz:int, eps=1e-5, affine=True, track_running_stats=True, momentum=0.1):
-        self.eps, self.track_running_stats, self.momentum = eps, track_running_stats, momentum
-        if affine: self.weight, self.bias = Tensor.ones(sz, dtype=dtypes.float32), Tensor.zeros(sz, dtype=dtypes.float32)
-        else: self.weight, self.bias = None, None
-
-        self.running_mean, self.running_var = Tensor.zeros(sz, requires_grad=False, dtype=dtypes.float32), Tensor.ones(sz, requires_grad=False, dtype=dtypes.float32)
-        self.num_batches_tracked = Tensor.zeros(1, requires_grad=False, dtype=dtypes.int)
-      def __call__(self, x:Tensor):
-        batch_mean = self.running_mean
-        batch_invstd = self.running_var.reshape(1, -1, 1, 1).expand(x.shape).add(self.eps).rsqrt()
-        return x.batchnorm(self.weight, self.bias, batch_mean, batch_invstd).cast(dtypes.default_float)
-    resnet.BatchNorm = functools.partial(CustBatchNorm, eps=0.0)
-  else:
-    class CustUnsyncedBN(UnsyncedBatchNorm):
-      def __call__(self, x:Tensor):
-        if isinstance(x.lazydata, MultiLazyBuffer): assert x.lazydata.axis is None or x.lazydata.axis == 0 and len(x.lazydata.lbs) == self.num_devices
-        nd = x.shape[0]%self.num_devices
-        if nd==0: nd = self.num_devices
-        xr = x.reshape(nd, -1, *x.shape[1:]).cast(dtypes.float32)
-        batch_mean, batch_invstd = self.calc_stats(xr)
-        ret = xr.batchnorm(
-          self.weight.reshape(1, -1).expand((nd, -1)),
-          self.bias.reshape(1, -1).expand((nd, -1)),
-          batch_mean, batch_invstd, axis=(0, 2))
-        return ret.reshape(x.shape).cast(x.dtype)
-    resnet.BatchNorm = functools.partial(CustUnsyncedBN, num_devices=len(GPUS))
+  if SYNCBN: resnet.BatchNorm = functools.partial(BatchNormRetinanet, eps=0.0)
+  else: resnet.BatchNorm = functools.partial(UnSyncedBatchNormRetinanet, num_devices=len(GPUS))
   resnet_model = resnet.ResNeXt50_32X4D()
   resnet_model.load_from_pretrained()
